PWGEM/PhotonMeson: fix MC indexing for daughters (#5600)

+++ PWGEM/PhotonMeson/Core/HistogramsLibrary.cxx
@@ -403,12 +403,12 @@ void o2::aod::pwgem::photon::histogram::DefineHistograms(THashList* list, const 
   }
 
   const int nmgg = 401;
-  float mgg[nmgg] = {};
+  double mgg[nmgg] = {};
   for (int i = 0; i < nmgg; i++) {
     mgg[i] = 0.002 * i;
   }
   const int npTgg = 71;
-  float pTgg[npTgg] = {};
+  double pTgg[npTgg] = {};
   for (int i = 0; i < 50; i++) {
     pTgg[i] = 0.1 * (i - 0) + 0.0; // from 0 to 5 GeV/c, every 0.1 GeV/c
   }
@@ -527,12 +527,12 @@ void o2::aod::pwgem::photon::histogram::DefineHistograms(THashList* list, const 
   }
 
   const int nmgg04 = 201;
-  float mgg04[nmgg04] = {};
+  double mgg04[nmgg04] = {};
   for (int i = 0; i < nmgg04; i++) {
     mgg04[i] = 0.002 * i;
   }
   const int npTgg10 = 61;
-  float pTgg10[npTgg10] = {};
+  double pTgg10[npTgg10] = {};
   for (int i = 0; i < 50; i++) {
     pTgg10[i] = 0.1 * (i - 0) + 0.0; // from 0 to 5 GeV/c, every 0.1 GeV/c
   }
